Move commander to its own module
